refactor: upgrader file renaming
