+++ packages/nocodb/src/version-upgrader/upgraders/0104002_ncAttachmentUpgrader.ts
