+++ packages/nocodb/src/version-upgrader/upgraders/0105002_ncStickyColumnUpgrader.ts
