+++ packages/nocodb/src/version-upgrader/upgraders/0107004_ncProjectConfigUpgrader.ts
